Atualização

+++ src/Fonte/ProdutoUnidade.java
@@ -33,8 +33,8 @@ public class ProdutoUnidade extends Produto{
         return quantidade;
     }
 
-    public void setQuantidade(int quantidade) {
-        this.quantidade = quantidade;
+    public void setQuantidade(double quantidade) {
+        this.quantidade = (int) quantidade;
     }
 
     /**
